% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_table_trips.R
\name{batch_table_trips}
\alias{batch_table_trips}
\title{Make a table of payments from a batch of pay stubs}
\usage{
batch_table_trips(pdf_stub_list)
}
\arguments{
\item{pdf_stub_list}{the pdf files with the payment info}
}
\description{
This function takes pdfs of pay statement that are downloaded from uber
  https://partners.uber.com/p3/money/statements/all
  The settings must correspond to: portrait mode (i.e. wide), paper size A4,
  margins none. It returns a table of the trip data.
}
\examples{
batch_table_trips()
}
\keyword{accounting}
\keyword{pdf}
\keyword{uber}
