% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebit_tex.R
\name{ebit_tex}
\alias{ebit_tex}
\title{Make a latex table of the income statement}
\usage{
ebit_tex(ebit, year = NA)
}
\arguments{
\item{ebit}{EBIT table generated by uber_ebit()}

\item{year}{Default to current year}
}
\description{
This function takes a dataframe of uber income data and user generated
  expsenses and combines them to make an income statement up to EBIT.
  The statement can be either quarterly or annual (10k or 10q).
}
\examples{
ebit_tex( my10q, year = 2015)
}
\keyword{EBIT}
\keyword{accounting}
\keyword{latex}
\keyword{uber}
