% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_inc.R
\name{fmt_inc}
\alias{fmt_inc}
\title{Make an income statement}
\usage{
fmt_inc(income_data, tenQ = FALSE)
}
\arguments{
\item{income_data}{dataframe of income data generated by table trips}

\item{tenQ}{should the table be quarterly or annual ? Default to annual}
}
\description{
This function takes the delivery data and makes an income statement.
  Optionally, an expense sheet can be included. The statement can be
  either quarterly or annual (10k or 10q).
}
\examples{
fmt_inc( mydeliveries.df, tenQ = TRUE)
}
\keyword{accounting}
\keyword{uber}
