% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_trips.R
\name{table_trips}
\alias{table_trips}
\title{Read pdf pay statement and return trip of tables}
\usage{
table_trips(pdf_stub)
}
\arguments{
\item{pdf_stub}{the pdf file with the payment info}
}
\description{
This function takes a pdf pay statement that is downloaded from uber
  https://partners.uber.com/p3/money/statements/all
  The settings must correspond to: portrait mode (i.e. wide), paper size A4,
  margins none. It returns a table of the trip data.
}
\examples{
table_trips()
}
\keyword{accounting}
\keyword{pdf}
\keyword{uber}
