% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uber_ebit.R
\name{uber_ebit}
\alias{uber_ebit}
\title{Make an income statement upto EBIT for Uber income and expense data}
\usage{
uber_ebit(income_data, expense_data, tenQ = FALSE, tex = FALSE)
}
\arguments{
\item{income_data}{dataframe of income generated by fmt_inc()}

\item{expense_data}{dataframe of expenses generated by fmt_exp()}

\item{tenQ}{should the table be quarterly or annual ? Default to annual}

\item{tex}{if TRUE, prints a latex formatted table as well.}
}
\description{
This function takes a dataframe of uber income data and user generated
  expsenses and combines them to make an income statement up to EBIT.
  The statement can be either quarterly or annual (10k or 10q).
}
\examples{
uber_ebit( myinc.df, myexp.df, tenQ = TRUE, tex = FALSE)
}
\keyword{EBIT}
\keyword{accounting}
\keyword{uber}
